break meetings flow into sections
