+++ app/frontend/components/domains/project-meeting/project-meeting-screen/sections/review-summary-section.tsx
@@ -1,0 +1,29 @@
+import { Box, Heading, Link } from "@chakra-ui/react"
+import React from "react"
+import { useTranslation } from "react-i18next"
+import { SummarySectionProps } from "../shared/types"
+
+interface ReviewSummarySectionProps extends SummarySectionProps {
+  onNavigateToSection: (sectionKey: string) => void
+}
+
+export const ReviewSummarySection = ({
+  title,
+  sectionKey,
+  children,
+  onNavigateToSection,
+}: ReviewSummarySectionProps) => {
+  const { t } = useTranslation()
+
+  return (
+    <Box mb={8}>
+      <Heading as="h2" size="lg" variant="yellowline" mb={4}>
+        {title}
+      </Heading>
+      {children}
+      <Link as="button" type="button" color="text.link" mt={2} onClick={() => onNavigateToSection(sectionKey)}>
+        {t("ui.change")}
+      </Link>
+    </Box>
+  )
+}
