+++ app/frontend/components/domains/project-meeting/project-meeting-screen/shared/constants.ts
@@ -1,0 +1,1 @@
+export const ACCEPTED_DOCUMENT_TYPES = ["application/pdf", ".pdf", ".jpg", ".jpeg", ".png", ".doc", ".docx"]
