+++ app/frontend/components/domains/project-meeting/project-meeting-screen/shared/document-utils.ts
@@ -1,0 +1,10 @@
+import { EMeetingRequestDocumentType } from "../../../../../types/enums"
+import { MeetingRequestDocumentFormValue } from "./types"
+
+const getDocumentType = (document: MeetingRequestDocumentFormValue) =>
+  document.documentType || EMeetingRequestDocumentType.supporting
+
+export const activeDocumentsForType = (
+  documents: MeetingRequestDocumentFormValue[],
+  documentType: EMeetingRequestDocumentType
+) => documents.filter((document) => !document._destroy && getDocumentType(document) === documentType)
