+++ app/frontend/components/domains/project-meeting/project-meeting-screen/shared/types.ts
@@ -1,0 +1,10 @@
+import React from "react"
+import { IMeetingRequestDocument } from "../../../../../types/types"
+
+export type MeetingRequestDocumentFormValue = Partial<IMeetingRequestDocument>
+
+export interface SummarySectionProps {
+  title: string
+  sectionKey: string
+  children: React.ReactNode
+}
